requirements.txt
streamlit
pandas
streamlit-authenticator
PyYAML
google-generativeai